% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_file_ext.R
\name{assemble_file_ext}
\alias{assemble_file_ext}
\title{assemble_file_ext.R}
\usage{
assemble_file_ext(metadata_resources, lovs = ddhconnect::get_lovs())
}
\arguments{
\item{metadata_resources}{list: object returned by get_metadata_resources}

\item{lovs}{dataframe: object returned by the get_lovs() function}
}
\value{
dataframe
}
\description{
Gather the file extension and the field_form value
}
