% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_embargo_date.R
\name{check_embargo_date}
\alias{check_embargo_date}
\title{check_embargo_date}
\usage{
check_embargo_date(metadata_dataset)
}
\arguments{
\item{metadata_dataset}{list: object returned by get_metadata() on a dataset}

\item{lovs}{dataframe: object returned by the get_lovs() function}
}
\value{
list
}
\description{
Check if the dataset's embargo date has passed
}
