% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_file_path.R
\name{check_file_path}
\alias{check_file_path}
\title{check_file_path}
\usage{
check_file_path(metadata_resource)
}
\arguments{
\item{metadata_resource}{list: output of get_metadata() for a resource}
}
\value{
character vector
}
\description{
Check if the resources have the correct path based on data classification
}
