% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing.R
\name{check_missing}
\alias{check_missing}
\title{check_missing}
\usage{
check_missing(metadata_dataset)
}
\arguments{
\item{metadata_dataset}{list: output of get_metadata() for a dataset}
}
\value{
character vector
}
\description{
Get missing resources for all datasets
}
