% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_overall.R
\name{check_overall}
\alias{check_overall}
\title{check_overall}
\usage{
check_overall(nid_dataset, recom_fields = TRUE, file_ext = TRUE,
  file_links = TRUE, credentials = list(cookie = dkanr::get_cookie(), token
  = dkanr::get_token()))
}
\arguments{
\item{metadata_datasets}{list: object returned by get_metadata_datasets}

\item{lovs}{dataframe: object returned by the get_lovs() function}
}
\value{
dataframe
}
\description{
Overall automated quality checks for the Data Catalog
}
