% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_privacy_status.R
\name{check_privacy_status}
\alias{check_privacy_status}
\title{check_privacy_status}
\usage{
check_privacy_status(metadata_resource, lovs = ddhconnect::get_lovs())
}
\arguments{
\item{metadata_resource}{list: output of get_metadata() for a resource}
}
\value{
character vector
}
\description{
Check if the confidential/official use only dataset has public resources
}
