% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_recom_fields.R
\name{check_recom_fields}
\alias{check_recom_fields}
\title{check_recom_fields}
\usage{
check_recom_fields(metadata_datasets, credentials = list(cookie =
  dkanr::get_cookie(), token = dkanr::get_token()))
}
\arguments{
\item{metadata_datasets}{list: object returned by get_metadata_datasets}

\item{lovs}{dataframe: object returned by the get_lovs() function}
}
\value{
dataframe
}
\description{
Check if the resource file extension matches the field_form value
}
