% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_resource_link.R
\name{check_resource_link}
\alias{check_resource_link}
\title{check_resource_link}
\usage{
check_resource_link(metadata_resource)
}
\arguments{
\item{metadata_resources}{list: object returned by get_metadata}
}
\value{
dataframe
}
\description{
Check for 404 errors in resource links
}
