% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_resource_links.R
\name{check_resource_links}
\alias{check_resource_links}
\title{check_resource_links}
\usage{
check_resource_links(metadata_resources)
}
\arguments{
\item{metadata_resources}{list: object returned by get_metadata_resources or get_metadata_all}
}
\value{
dataframe
}
\description{
Check for 404 errors in resource links
}
