% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_unpublished.R
\name{check_unpublished}
\alias{check_unpublished}
\title{check_unpublished}
\usage{
check_unpublished(metadata_resource)
}
\arguments{
\item{metadata_resource}{list: output of get_metadata() for a resource}
}
\value{
character vector
}
\description{
Check if resource is unpublished
}
