% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_machine_names.R
\name{count_machine_names}
\alias{count_machine_names}
\title{count_machine_names}
\usage{
count_machine_names(all_data_fields)
}
\arguments{
\item{all_metadata}{list: object returned by get_all_metadata}

\item{credentials}{list: object returned by the get_credentials() function}
}
\value{
dataframe
}
\description{
View the number of times a machine name is used in the metadata. Keeps count of used and missing
}
