% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata_datasets.R
\name{get_metadata_datasets}
\alias{get_metadata_datasets}
\title{get_metadata_datasets}
\usage{
get_metadata_datasets(nids_datasets, credentials = list(cookie =
  dkanr::get_cookie(), token = dkanr::get_token()))
}
\arguments{
\item{nids_datasets}{list: list of nodes to gather data for}

\item{credentials}{list: object returned by the get_credentials() function}
}
\value{
list
}
\description{
Gather all the metadata corresponding to every nid for datasets
}
