% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata_resources.R
\name{get_metadata_resources}
\alias{get_metadata_resources}
\title{get_metadata_resources}
\usage{
get_metadata_resources(metadata_datasets, credentials = list(cookie =
  dkanr::get_cookie(), token = dkanr::get_token()))
}
\arguments{
\item{credentials}{list: object returned by the get_credentials() function}

\item{nids_datasets}{list: list of nids for resources}
}
\value{
list
}
\description{
Gather all the metadata corresponding to every nid for resource
}
