% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_resources_list.R
\name{get_resources_list}
\alias{get_resources_list}
\title{get_resources_list()}
\usage{
get_resources_list(datasets_metadata, rand_count = 1e+05,
  credentials = list(cookie = dkanr::get_cookie(), token =
  dkanr::get_token()))
}
\arguments{
\item{datasets_metadata}{character: output of get_metadata_datasets}

\item{credentials}{list: API authentication credentials}

\item{est_count}{character: Approximate number of resources}
}
\value{
numeric vector
}
\description{
get_resources_list()
}
