% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unpublished_resource_nids.R
\name{get_unpublished_resource_nids}
\alias{get_unpublished_resource_nids}
\title{get_unpublished_resources}
\usage{
get_unpublished_resource_nids(datatype = "Microdata")
}
\arguments{
\item{datatype}{character: datatype of datasets for which to return the resource nids}
}
\value{
character vector
}
\description{
Get unpublished resources of one or all of the data types
}
