% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_file_ext.R
\name{missing_file_ext}
\alias{missing_file_ext}
\title{missing_file_ext.R}
\usage{
missing_file_ext(df)
}
\arguments{
\item{df}{list: object returned by get_metadata_resources}

\item{credentials}{list: object returned by the get_credentials() function}
}
\value{
dataframe
}
\description{
Check if the resource file extension matches the field_form value
}
