% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish_resources.R
\name{publish_resources}
\alias{publish_resources}
\title{publish_resources}
\usage{
publish_resources(resource_nids)
}
\arguments{
\item{resource_nids}{character vector: list of nodes to gather data for}
}
\value{
character
}
\description{
Publish resources from the list of node ids
}
