% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_overall.R
\name{check_overall}
\alias{check_overall}
\title{check_overall}
\usage{
check_overall(nid_data, checks = c("recommended", "exts", "links",
  "unpublished"), credentials = list(cookie = dkanr::get_cookie(), token =
  dkanr::get_token()))
}
\arguments{
\item{nid_data}{string: node id for a dataset}

\item{lovs}{dataframe: object returned by the get_lovs() function}
}
\value{
dataframe
}
\description{
Overall automated quality checks for the Data Catalog
}
