% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_recom_fields.R
\name{check_recom_fields}
\alias{check_recom_fields}
\title{check_recom_fields}
\usage{
check_recom_fields(dataset)
}
\arguments{
\item{datasets}{list: object returned by get_metadata()}

\item{lovs}{dataframe: object returned by the get_lovs() function}
}
\value{
vector
}
\description{
Check which recommended fields are missing
}
