% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_resources_list.R
\name{get_resources_list}
\alias{get_resources_list}
\title{get_resources_list()}
\usage{
get_resources_list(est_count = 1e+05, root_url = dkanr::get_url(),
  credentials = list(cookie = dkanr::get_cookie(), token =
  dkanr::get_token()))
}
\arguments{
\item{est_count}{character: Approximate number of resources}

\item{credentials}{list: API authentication credentials}
}
\value{
numeric vector
}
\description{
get_resources_list()
}
